First excel upload project
The idea behind the data, questions, and answers is a project I was working on, I won't mention the name bcoz of the nature and the sensitivity of the data. But I was working on the data to determine the impression, clicks on the website and the orders made in an e-commerce platform.